ASCII85 encoding (stand alone file)
I can't take credit for all the code here, it's a mash up of two sources, ported from C and I believe PHP.  It's the encoding standard you commonly see in PDF's, and this should be compatible with that if you remove the 'y' character code.  It is not compatible that I am aware of with RFC 1924 (that uses a different character set).